Damageの処理を変更・Cursorをクラス化

+++ プロジェクト一式/Scene/ResultScene.cpp
@@ -39,7 +39,7 @@ void ResultScene::Initialize()
 		po->SetRotateY(230.0f);
 		InstantiateFront<WeaponMainOrnament>(po);
 		Model::SetAnimFrame(po->GetModelHandle(), 
-			PLAYER_ANIMATION_DATA[PLAYER_ANIMATION::DEAD].endFrame, PLAYER_ANIMATION_DATA[PLAYER_ANIMATION::DEAD].endFrame, 1.0f);
+			PLAYER_ANIMATION_DATA[(int)PLAYER_ANIMATION::DEAD].endFrame, PLAYER_ANIMATION_DATA[(int)PLAYER_ANIMATION::DEAD].endFrame, 1.0f);
 	}
 	else {
 		Camera::SetPosition(XMFLOAT3(15, 1, 15));
@@ -49,7 +49,7 @@ void ResultScene::Initialize()
 		po->SetRotateY(230.0f);
 		InstantiateFront<WeaponMainOrnament>(po);
 		Model::SetAnimFrame(po->GetModelHandle(), 
-			PLAYER_ANIMATION_DATA[PLAYER_ANIMATION::IDLE].startFrame, PLAYER_ANIMATION_DATA[PLAYER_ANIMATION::IDLE].endFrame, 1.0f);
+			PLAYER_ANIMATION_DATA[(int)PLAYER_ANIMATION::IDLE].startFrame, PLAYER_ANIMATION_DATA[(int)PLAYER_ANIMATION::IDLE].endFrame, 1.0f);
 	}
 
 }
